Move saving functionality to `numpy_dataset_utils`.
CONTEXT: `numpy_dataset_utils` is a general package for writing numpy datasets (which will be used to store scatterer distributions) to disk.
